1.1-alpha.5

+++ src/main/java/ru/massonnn/masutils/client/config/Mineshaft.java
@@ -31,5 +31,5 @@ public class Mineshaft {
 
     public boolean mineshaftCommands = false;
 
-    public int traceThickness = 4;
+    public float traceThickness = 4f;
 }
